Work in progress
- Work towards getting `SchemaProvider`s integrated into the system as per issue #321
- SparkBackend code has been split between:
   - `mimir.data.SparkSchemaProvider`: Access to spark's Derby/Hive backend data repository
   - `mimir.exec.spark.*` and `mimir.exec.Compiler`: Code related to Spark compilation and execution management.
- Database no longer takes a backend parameter (just a metadata backend).  Spark is now the only query processing infrastructure used.
- Code throughout has been re-factored to manage a global spark context (TODO: Further refactor to make the global spark context a feature of Database)
- OperatorTranslation has been renamed to `RAToSpark` for consistency with the other translation classes.
