+++ src/main/scala/mimir/exec/mode/BestGuess.scala
@@ -84,7 +84,7 @@ object BestGuess
 
 
     // Replace VG-Terms with their "Best Guess values"
-    oper = bestGuessQuery(db, oper)
+    oper = InlineVGTerms(oper, db)
 
     logger.debug(s"GUESSED: $oper")
 
@@ -129,28 +129,6 @@ object BestGuess
       operRaw.columnNames,
       provenanceCols
     )
-  }
-
-  /**
-   * Remove all VGTerms in the query and replace them with the 
-   * equivalent best guess values
-   */
-  def bestGuessQuery(db: Database, oper: Operator): Operator =
-  {
-    // Remove any VG Terms for which static best-guesses are possible
-    // In other words, best guesses that don't depend on which row we're
-    // looking at (like the Type Inference or Schema Matching lenses)
-    val mostlyDeterministicOper =
-      InlineVGTerms(oper, db)
-
-    // Deal with the remaining VG-Terms.  
-    if(db.backend.canHandleVGTerms){
-      // The best way to do this would be a database-specific "BestGuess" 
-      // UDF if it's available.
-      return mostlyDeterministicOper
-    } else {
-      throw new RAException("Error, Best Guess Cache Doesn't Work")
-    }
   }
 
   def wrap(db: Database, results: ResultIterator, query: Operator, meta: MetadataT): ResultIterator =
